first attempt

+++ public/js/buy.js
@@ -6,27 +6,71 @@ $(document).ready(function () {
 
   getPlants();
 
-  // $("#searchBarFilt").on("keyup", function() {
-  //   var value = $(this).val().toLowerCase();
-  //   $(".card").filter(function() {
-  //     $(this).toggle($(this).text().toLowerCase().indexOf(value) > -1)
-  //     });
-  // });
-
- 
-
 // If purchase btn clicked
 $(document).on("click",".buy-btn", function(){
 
-  var id = this.id;
-  console.log(id);
+  var id = this.id
 
-  $.ajax({
-    method: "DELETE",
-    url: "/api/buy/" + id
-  }).then(getPlants);
+  getPrice(id)
 
 });
+
+// function to get price of the selected plant
+function getPrice(id){
+
+  $.get("/api/price/" +id).then(function(data) {
+    var newPrice = parseInt(data.price);
+    getWallet(newPrice,id);
+    console.log(newPrice);
+  });
+}
+
+// function to get the current balance in the user wallet
+function getWallet(newPrice,id){
+  $.post("/api/wallet").then(function(data) {
+    var balance = parseInt(data.wallet);
+    console.log(balance);
+    checkout(newPrice,balance,id);
+  });
+}
+
+
+function checkout(newPrice,balance,id){
+
+  //able to buy
+  if (balance >= newPrice){
+    var newBalance = balance - newPrice;
+    updateWallet(newBalance);
+
+    $.ajax({
+    method: "DELETE",
+    url: "/api/buy/" + id
+    }).then(getPlants);
+
+
+  }
+  //unable to buy
+  else{
+    console.log("Not enough funds");
+  }
+}
+
+//function to update wallet
+function updateWallet(newBalance) {
+
+  console.log(newBalance)
+
+  $.ajax({
+    method: "PUT",
+    url: "/api/wallet",
+    data: {
+      wallet: newBalance}
+  })
+
+}
+
+
+
 
 
 
@@ -40,7 +84,7 @@ function getPlants() {
 
     for (var a = 0; a < data.length; a++) {
 
-      var templateString = '<h2>' + (Object.values(data[a].plantName).join('')) + '</h2><h4> $ ' + Object.values(data[a].price).join('') + '</h4><p>' + Object.values(data[a].description).join('');
+      var templateString = '<h2>' + (Object.values(data[a].plantName).join('')) + '</h2><h4 id = "pricetag"> $ ' + Object.values(data[a].price).join('') + '</h4><p>' + Object.values(data[a].description).join('');
       var newImg = (Object.values(data[a].imgURL).join(''));
       var newId = data[a].id;
